Refactored to move most of behavior into parent Algorithm Class
